+++ terraform/envs/stage/providers.tf
@@ -16,6 +16,10 @@ terraform {
       source  = "hashicorp/helm"
       version = "~> 2"
     }
+    tls = {
+      source  = "hashicorp/tls"
+      version = "~> 4"
+    }
   }
 }