+++ terraform/modules/k8s-ingress/versions.tf
@@ -1,0 +1,17 @@
+terraform {
+  required_version = ">= 1.6.0"
+  required_providers {
+    kubernetes = {
+      source  = "hashicorp/kubernetes"
+      version = "~> 2"
+    }
+    helm = {
+      source  = "hashicorp/helm"
+      version = "~> 2"
+    }
+    tls = {
+      source  = "hashicorp/tls"
+      version = "~> 4"
+    }
+  }
+}